+++ CHANGELOG.md
@@ -19,6 +19,7 @@ This changelog only contains the changes that are unreleased. For changes for in
 - Updates for instances not showing the button until a restart [#677]
 - Issue with instance settings allowing setting of memory java params
 - Adding mods from CurseForge not filtering out Quilt files on non Quilt instances
+- Auto select latest file when updating mods
 
 ### Misc
 - Implement view model for InstancesTab [#675]